requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
requests
openai